Rename baserow.ts -> baserow.app.ts
