+++ components/baserow/actions/common.ts
@@ -1,4 +1,4 @@
-import baserow from "../app/baserow";
+import baserow from "../app/baserow.app";
 export default {
   props: {
     baserow,
