+++ components/baserow/actions/get-row/get-row.ts
@@ -1,5 +1,5 @@
 import { defineAction } from "@pipedream/types";
-import baserow from "../../app/baserow";
+import baserow from "../../app/baserow.app";
 import { DOCS_LINK } from "../../common/constants";
 import {
   GetRowParams, Row,
